refactor: 홈 초기 팔로잉 추천과 덕질고사 추천 분리
